Parse for Debugger written
At this point, the debugger can read in a file, and pull all pertinent data out of it
It keeps track of all labels, and it separates lines of content with lines stored in memory.

I realized after writing the last thing, that I didn't implement multiplication yet.
It was making my brain hurt. I'll go back for it in a bit.

On limitation, is that since it doesn't compile, yet, the instructions themselves are not kept in memory, and the file can't dynamically modify itself.

Lis should work though, so long as it's a .word line that follows it.