Works
Ok. So I've got a working debugger.
Actually, it's more operating than working.
I haven't tested everything yet

Also, it isn't terribly useful for anything right now other than seeing where errors occur.

I cheated on mult and multu. I just multiplied it out and set it MFHI to 0. This means that if you multiply large numbers together, you get an overflow in MFLO, and MFHI is still empty.

This should work out ok for my purposes, though.

All that's needed to make it more useful is a run to line function, and a viewer of registers and memory on the fly.